Cortex_OS/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
litellm==1.17.9
openai==1.3.7
anthropic==0.7.8
groq==0.4.1
qdrant-client==1.7.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyYAML==6.0.1
Jinja2==3.1.2
cryptography==41.0.8
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0